refactor(sdk): move generator package to apps/sdk-generator
