+++ apps/sdk-generator/ownership.ts
