+++ apps/sdk-generator/package.json
