+++ apps/sdk-generator/tests/ownership.test.ts
