+++ apps/sdk-generator/tests/paths.test.ts
