+++ package.json
@@ -7,7 +7,7 @@
 			"packages/debugger",
 			"packages/types",
 			"packages/api",
-			"packages/api/scripts/generate",
+			"apps/sdk-generator",
 			"packages/plugin-test",
 			"packages/utils",
 			"packages/logger",
@@ -30,7 +30,7 @@
 	},
 	"scripts": {
 		"prepare": "lefthook install --reset-hooks-path",
-		"generate-sdk": "bun packages/api/scripts/generate/index.ts",
+		"generate-sdk": "bun apps/sdk-generator/index.ts",
 		"lint": "oxlint --config .oxlintrc.json apps packages",
 		"lint:fix": "oxlint --config .oxlintrc.json --fix-dangerously apps packages",
 		"fmt": "oxfmt --config .oxfmtrc.json apps packages",
